+++ changelog.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fixed double slash in simplified URL for some permalink settings
+- Fixed wrong hidden language in language switcher
 
 ## [2.9.0] - 23.09.2025